address comments v2 - merge it tests and add default values for configs
